+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - [Backup NG logs] Add the job's name to the modal's title [#2711](https://github.com/vatesfr/xen-orchestra/issues/2711) (PR [#3115](https://github.com/vatesfr/xen-orchestra/pull/3115))
 - Adding a XCP-ng host to a XS pool now fails fast [#3061](https://github.com/vatesfr/xen-orchestra/issues/3061) (PR [#3118](https://github.com/vatesfr/xen-orchestra/pull/3118))
 - [Backup NG logs] Ability to report a failed job and copy its log to the clipboard [#3100](https://github.com/vatesfr/xen-orchestra/issues/3100) (PR [#3110](https://github.com/vatesfr/xen-orchestra/pull/3110))
+- [Backup NG form] Make copy retention optional [#2711](https://github.com/vatesfr/xen-orchestra/issues/2711)(PR [3113](https://github.com/vatesfr/xen-orchestra/pull/3113))
 
 ### Bug fixes